Conflicts:
	arangod/Replication/Syncer.cpp
	arangod/VocBase/index-garbage-collector.c
	build_posix.h
	html/admin/api-docs/traversal.json
	js/apps/aardvark/api-docs/collection.json
	js/apps/aardvark/api-docs/cursor.json
	js/apps/aardvark/api-docs/document.json
	js/apps/aardvark/api-docs/edge.json
	js/apps/aardvark/api-docs/edges.json
	js/apps/aardvark/api-docs/graph.json
	js/apps/aardvark/api-docs/index.json
	js/apps/aardvark/api-docs/replication.json
	js/apps/aardvark/api-docs/simple.json
	js/apps/aardvark/api-docs/system.json

+++ arangod/FulltextIndex/fulltext-list.c
@@ -622,7 +622,7 @@ TRI_fulltext_list_t* TRI_InsertListFulltextIndex (TRI_fulltext_list_t* list,
     TRI_fulltext_list_t* clone;
     uint32_t newSize;
 
-    newSize = numEntries * GROWTH_FACTOR;
+    newSize = (uint32_t) (numEntries * GROWTH_FACTOR);
     
     if (newSize == numEntries) {
       // 0 * something might not be enough...
